- Fixed usrloc part - Added support for jabber module - Changed register_watcher and unregister_watcher prototypes, the function   have now From, To, callback and opaque data as parameters. - Small bug fixes - Removed sl module dependency - t_reply is now used for replies which means that subscribe must be called   from if (t_newtran())... block ! - Dialog support still tricky, needs to be rewritten - kphone is unable to process our XML document even if it is correct, I have   to see kphone sources to see why

+++ modules_s/pa/watcher.c
@@ -116,7 +116,7 @@ void print_watcher(FILE* _f, watcher_t* _w)
 	fprintf(_f, "~~~Watcher~~~\n");
 	fprintf(_f, "from   : \'%.*s\'\n", _w->from.len, _w->from.s);
 	fprintf(_f, "contact: \'%.*s\'\n", _w->contact.len, _w->contact.s);
-	fprintf(_f, "expires: %d\n", _w->expires - time(0));
+	fprintf(_f, "expires: %d\n", (int)(_w->expires - time(0)));
 	fprintf(_f, "accept : %s\n", (_w->accept == DOC_XPIDF) ? ("DOC_XPIDF") : ("DOC_LPIDF"));
 	fprintf(_f, "next   : %p\n", _w->next);
 	fprintf(_f, "dialog.callid: \'%.*s\'\n", _w->dialog.callid.len, _w->dialog.callid.s);
